refactor: interface file division
